[FIX] UBLE-176 현지도 검색 버튼 로직 수정 및 매장 상세 카드 변경 (#220)
* UBLE-176 refactor: 훅 제거

* UBLE-176 refactor: 현 지도에서 검색 버튼 표시 로직 수정

* UBLE-176 refactor: 줌 레벨 매장명 표시 조건 분기 경우에만 업데이트

* UBLE-176 refactor: memo 이용하여 Skeleton 불편함 제거

* UBLE-176 fix: conflict 해결

* UBLE-176 refactor: 의존성 배열 변경

* UBLE-176 refactor: 아이콘 업데이트 조건 상단 이동

* UBLE-176 refactor: 미사용 import 제거
